% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_pvals.R
\name{comb_pvals}
\alias{comb_pvals}
\title{Fisher's rule for combining several p-values}
\usage{
comb_pvals(pv)
}
\arguments{
\item{pv}{the vector of pvalues to be combiend together}
}
\value{
the Fisher combination of the p-values. See Details.
}
\description{
According to Fisher's rule, if the p-values are correlated, then this does not follow a simple chi-square 
mixture under the null.
}

