% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{common1Mean_C}
\alias{common1Mean_C}
\title{Fast C++ implementation of mean number of 1's in common for a pair of rows, each taken from one 
of 2 binary matrices}
\usage{
common1Mean_C(mat_A, mat_B)
}
\arguments{
\item{mat_A}{a \code{nB x K} matrix of the observations to be matched. Must be positive integers 
(supposed to be binary).}

\item{mat_B}{a \code{nA x K} matrix of the database into which a match is looked for. Must be 
positive integers (supposed to be binary).}
}
\description{
Fast C++ implementation of mean number of 1's in common for a pair of rows, each taken from one 
of 2 binary matrices
}
\examples{
mat1 <- matrix(round(rnorm(n=500, sd=1.2)), ncol=10, nrow=50)
mat2 <- rbind(mat1[1:10, ],
             matrix(round(rnorm(n=400, sd=1.2)), ncol=10, nrow=40)
             )
rownames(mat1) <- paste0("A", 1:nrow(mat1))
rownames(mat1) <- paste0("B", 1:nrow(mat1))
mat1 <- 1*(mat1>1)
mat2 <- 1*(mat2>1)
common1Mean_C(mat1, mat2)
m <- numeric(50^2)
for(i in 1:50){
for(j in 1:50){
  m[(i-1)*50 + j] <- sum(mat1[i, ]*mat2[j, ])
}
}
mean(m)

}

