% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dateread}
\alias{dateread}
\title{C++ function for reading dates from character stings}
\usage{
dateread(s, fmt)
}
\arguments{
\item{s}{a date as a character string}

\item{fmt}{a fomat to read the date}
}
\description{
C++ function for reading dates from character stings
}
\examples{
dateread(" Nov 10 2008 12:00AM", "\%B\%d\%Y")
dateread(" Nov 10 2008 12:00AM", " \%B  \%d \%Y")
dateread(" Nov 10 2008 12:00AM", " \%d \%B \%Y")

}
\keyword{internal}

