% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{datetest}
\alias{datetest}
\title{C++ function for comparing dates encoded as character stings}
\usage{
datetest(s1, s2, fmt1, fmt2)
}
\arguments{
\item{s1}{a date as a character string}

\item{s2}{a second date as a character string}

\item{fmt1}{a fomat to read the date \code{s1}}

\item{fmt2}{a fomat to read the date \code{s2}}
}
\value{
\code{TRUE} if \code{s1} precedes \code{s2}, \code{FALSE} otherwise.
}
\description{
C++ function for comparing dates encoded as character stings
}
\examples{
dateread("20NOV2008 ", "\%d\%B\%Y")
dateread(" JAN 10 2008", "\%B \%d\% Y")
dateread(" Jan 10 2008 12:00AM", " \%B  \%d \%Y")>=dateread(" Jan 26 2008 12:00AM", " \%B  \%d \%Y")
datetest(" Nov 3 2008 12:00AM", "04NOV2008", " \%B \%d \%Y", "\%d\%B\%Y")
datetest(" Nov 5 2008 12:00AM", "04NOV2008", " \%B \%d \%Y", "\%d\%B\%Y")


}
\keyword{internal}

