% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_winkler.R
\name{em_winkler_big}
\alias{em_winkler_big}
\title{Implementation of Winkler's EM algorithm or Fellegi-Sunter matching method when the data are too 
big to compute the agreement matrix.}
\usage{
em_winkler_big(data1, data2, tol = 0.01, maxit = 50, do_plot = TRUE,
  verbose = FALSE)
}
\description{
Agreement is recomputed on the fly each time it is needed. The EM steps are completely done in C++
This decreases the RAM usage (still important though), at the cost of increasing computationnal time.
}

