% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_C_sparse_big}
\alias{EMstep_C_sparse_big}
\title{C++ implementation of the E and M steps from Winkler's EM algorithm estimating FS method 
using sparse matrices for big sample sizes}
\usage{
EMstep_C_sparse_big(mat_A, mat_B, p, m, u)
}
\description{
C++ implementation of the E and M steps from Winkler's EM algorithm estimating FS method 
using sparse matrices for big sample sizes
}
\keyword{internal}
