% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estep_C}
\alias{estep_C}
\title{C++ implementation of Winkler's Method E step}
\usage{
estep_C(agreemat, p, m, u, Log = FALSE)
}
\description{
C++ implementation of Winkler's Method E step
}
\keyword{internal}

