% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estep_C_sparse}
\alias{estep_C_sparse}
\title{C++ implementation of Winkler's Method E step
using a sparse agreement  matrix}
\usage{
estep_C_sparse(agreemat, p, m, u, Log = FALSE)
}
\description{
C++ implementation of Winkler's Method E step
using a sparse agreement  matrix
}
\keyword{internal}

