% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estep_C_vect}
\alias{estep_C_vect}
\title{Fast C++ implementation of Winkler's Method E step}
\usage{
estep_C_vect(agreemat, p, m, u)
}
\description{
Fast C++ implementation of Winkler's Method E step
}
\keyword{internal}
