% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estep_C_vect_sparse2}
\alias{estep_C_vect_sparse2}
\title{Fast C++ implementation of Winkler's Method E step
using two sparse matrices}
\usage{
estep_C_vect_sparse2(agreemat, agreemat_neg, p, m, u, Log = FALSE)
}
\description{
Saving memory while conserving speed by avoiding casting agreemat_neg
}
\keyword{internal}

