% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{expit}
\alias{expit_dev1}
\title{Logit link function}
\usage{
logit(p)

expit(x)

expit_dev1(x)
}
\arguments{
\item{p}{a real number in [0,1] to be transformed to the real scale 
( ]-infinity,infinity[ )}

\item{x}{a real number in ]-infinity,infinity[ to be transformed to the 
probability scale ( [0,1] )}

\item{x}{a real number in ]-infinity,infinity[ to be transformed to the 
probability scale ( [0,1] )}
}
\description{
inverse of the logistic transformation function

logistic transformation function

logistic transformation function
}
\keyword{internal}
