% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_cont2diff.R
\name{loglik_cont2diff}
\alias{loglik_cont2diff}
\title{Pseudo-likelihood computation}
\usage{
loglik_cont2diff(Amat, Bmat, d_max)
}
\arguments{
\item{Amat}{n x K0 matrix the database into which a match is looked for.}

\item{Bvec}{a vector of length K0 the observation to be matched.}

\item{eps_p}{}

\item{eps_n}{}

\item{piA}{}

\item{piB}{}
}
\description{
compute the loglikelihood of a match between a patient database and an observation, 
both of which having the same features measured.
}

