% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_FUN.R
\name{loglik_FUN}
\alias{loglik_FUN}
\title{Pseudo-likelihood computation}
\usage{
loglik_FUN(Amat, Bvec, eps_p, eps_n, piA, piB)
}
\arguments{
\item{Amat}{n x K0 matrix the database into which a match is looked for.}

\item{Bvec}{a vector of length K0 the observation to be matched.}

\item{eps_p}{a vector of length \code{K} giving the prior discrepancy rate 
expected from A to B for the positives, for each variable.}

\item{eps_n}{a vector of length \code{K} giving the prior discrepancy rate 
expected from A to B for the negatives, for each variable.}

\item{piA}{a vector of length \code{K} giving the prior probabilities of 
observing each variable in A.}

\item{piB}{a vector of length \code{K} giving the prior probabilities of 
observing each variable in B.}
}
\description{
Compute the loglikelihood of a match between a patient database and an observation, 
both of which having the same features measured.
}
\seealso{
loglikC
}

