% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikC}
\alias{loglikC}
\alias{loglikC_bin}
\alias{loglikC_bin_wDates}
\alias{loglikC_mat}
\alias{loglikC_mat_fast}
\alias{loglikratioC_diff}
\alias{loglikratioC_diff_arbitrary}
\title{C++ implementation of the pseudo-likelihood computation}
\usage{
loglikC(Bvec, Amat, eps_p, eps_n, piA, piB)

loglikC_mat(Bmat, Amat, eps_p, eps_n, piA, piB)

loglikC_mat_fast(Bmat, Amat, eps_p, eps_n, piA, piB)

loglikC_bin(Bmat, Amat, eps_p, eps_n, piA, piB)

loglikC_bin_wDates(Bmat, Amat, Bdates, Adates, eps_p, eps_n, piA, piB)

loglikratioC_diff(Bmat, Amat, d_max, eps_inf, pi_inf)

loglikratioC_diff_arbitrary(Bmat, Amat, d_max, cost)
}
\arguments{
\item{Bvec}{a vector of length K the observations to be matched.}

\item{Amat}{\code{nA x K} matrix the database into which a match is looked for.}

\item{eps_p}{a vector of length \code{K} giving the prior discrepancy rate 
expected from A to B for the positives, for each variable.}

\item{eps_n}{a vector of length \code{K} giving the prior discrepancy rate 
expected from A to B for the negatives, for each variable.}

\item{piA}{a vector of length \code{K} giving the prior probabilities of 
observing each variable in A.}

\item{piB}{a vector of length \code{K} giving the prior probabilities of 
observing each variable in B.}

\item{Bmat}{\code{K x nB} matrix of the observations to be matched.}

\item{Bdates}{\code{nB x K} matrix of the dates for each observations to be matched.}

\item{Adates}{\code{nA x K} matrix of the dates for database into which a match is looked for.}

\item{d_max}{a numeric vector of length \code{K} giving the minimum difference 
from which it is considered a discrepency.}

\item{eps_inf}{discrepancy rate for the differentiable variables}

\item{pi_inf}{the prior probability that the difference is greter than \code{d_max}}

\item{cost}{a numeric vector of length \code{K} giving the arbitrary cost of discrepency.}
}
\description{
C++ implementation of the pseudo-likelihood computation

\code{loglikC_mat_fast} is a faster C++ implementation of the pseudo-likelihood computation

\code{loglikC_bin} implements an even faster C++ implementation of the pseudo-likelihood computation for binary
variables

\code{loglikC_bin_wDates} implements a C++ implementation of the pseudo-likelihood computation for binary
variables with dates

\code{loglikratioC_diff} is a C++ implementation of the log-likelihood ratio computation for 
differentiable variables

Fast C++ implementation of the log-likelihood ratio computation for 
differentiating variables
}

