% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludic-package.R
\docType{package}
\name{ludic-package}
\alias{ludic}
\alias{ludic-package}
\title{ludic}
\description{
Linkage Using Diagnosis Codes
}
\details{
This package implements probabilistic record linkage methods that relies on the use of diagnosis 
codes only, in the absence of direct identifiers .

\tabular{ll}{
Package: \tab ludic\cr
Type: \tab Package\cr
Version: \tab ludic 0.1.3\cr
Date: \tab 2016-12-28\cr
License: \tab \href{https://cran.r-project.org/web/licenses/MIT}{The "MIT License" (MIT)}\cr
}
The main function of \code{ludic} is \link{recordLink}.
}
\author{
Boris P. Hejblum, Tianxi Cai
--- Maintainer: Boris P. Hejblum
}

