% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_2in1_bayes.R
\name{match_2in1_bayes}
\alias{match_2in1_bayes}
\title{Kernel for matching binary datasets by Probabilistic Patient Record Linkage}
\usage{
match_2in1_bayes(data, nb_rows = NULL, cut_bin = 0, MAF0 = 0.2,
  OR = 1.3, k0 = 5, nb_match = NULL, n1 = 3000, n2 = 2000,
  eps1 = 0.01, eps2 = eps1, bet_ptb = 6, type = "bin",
  aggreg_2ways = c("max_norm", "max", "min", "mean"),
  outputfile_name = "Default_output")
}
\arguments{
\item{data}{a matrix of original (continuous) dataset, either a matrix or a dataframe.}

\item{nb_rows}{the number of row to keep (equal or lower than the number of 
row of data). Default is \code{NULL} in which case all the rows are kept}

\item{cut_bin}{an integer specifying the threshold for dichotomizing observations in data.
Default is 0, in which case, any observation bigger than 0 becomes \code{TRUE} 
in the simulated data.}

\item{MAF0}{Minor Allele Frequency. Default is 0.2}

\item{OR}{Odds Ratio. Default is 1.3.}

\item{k0}{the number of top matching probabilities to store. Default is 5 top probabilities.}

\item{n1}{the number of patients in the first database. Default is 3000.}

\item{n2}{the number of patients in the second database. Default is 2000.}

\item{eps1}{probability that a feature is recorded in the first database but 
not in the second (mis-match or recording error). Default is 0.01.}

\item{eps2}{probability that a feature is recorded in the second database but 
not in the first (mis-match or recording error). Default is the same as \code{eps1}.}

\item{bet_ptb}{guessed value for the level of pertubation added}

\item{type}{the type of pertubation induced. Either \code{"bin"} for 
perturbation on the binary scale, or \code{"con"} for perturbation
induced on the continuous scale (before the phenotype variables are dichotomized).
Default is \code{"bin"}.}

\item{aggreg_2ways}{a charatcer string indicating how to merge the posterior two 
probability matrices obtained for each of the 2 databases. Four possibility are 
currently implemented: \code{"max_norm"}, \code{"max"}, \code{"min"} and \code{"mean"}.
Default is \code{"max_norm"}.}

\item{outputfile_name}{a character string for the name of the .txt file where the 
results should be stored. Must be provided without extension. Default is 'Default_output'.}

\item{nb_match.}{The number of true match to be simulated. Default is 3\% of the 
simulated pairs of patients.}
}
\value{
append the results in the file '\code{outputfile_name}.txt'.
}
\description{
Kernel for matching binary datasets by Probabilistic Patient Record Linkage
}

