% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchProbs_rank_multiprior.R
\name{matchProbs_rank_multiprior}
\alias{matchProbs_rank_multiprior}
\title{Compute the posterior probabilities for an observation to match other observations 
updating a matrix of prior matching probabilities}
\usage{
matchProbs_rank_multiprior(id0, computed_dist, match_priorprobs,
  yes_aggregate = FALSE, k0 = 5)
}
\arguments{
\item{id0}{the index for the reference observation in \code{1:n} to be match to p other observations}

\item{computed_dist}{an \code{n x p} matrix of computed distances used for ranking.}

\item{match_priorprobs}{an \code{n x p} matrix of prior matching probabilities.}

\item{yes_aggregate}{logical flag for wether the results are sorted according to the
original probability or the aggregated probability. Default is \code{TRUE}.}

\item{k0}{number of top (highest) matching probabilities to return. 
Default is 5.}
}
\value{
a character vector of length \code{k0*4} containing for each k0
potential match, inturn, the probability, the aggregated probability, the logLR 
abd the corresponding potential matching ID.
}
\description{
For a given observation of index id0 (in \code{1:n}), all the posterior 
matching probabilities are computed for the complementary \code{p} 
observations using a  matrix of prior matching probabilities previously calculated
}

