% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_pvals.R
\name{mp_pvals}
\alias{mp_pvals}
\title{P-values from matched probabilities}
\usage{
mp_pvals(data1, data2, ind_match, MAF0 = 0.2, OR = 1.3, k0 = 5,
  eps1 = 0.01, eps2 = eps1, aggreg_2ways = c("maxnorm", "max", "min",
  "mean", "prod"), nb_perturb = 1000, outputfile_name = "Default_output")
}
\arguments{
\item{MAF0}{Minor Allele Frequency. Default is 0.2}

\item{OR}{Odds Ratio. Default is 1.3.}

\item{k0}{the number of top matching probabilities to store. Default is 5 top probabilities.}

\item{eps1}{probability that a feature is recorded in the first database but 
not in the second (mis-match or recording error). Default is 0.01.}

\item{eps2}{probability that a feature is recorded in the second database but 
not in the first (mis-match or recording error). Default is the same as \code{eps1}.}

\item{aggreg_2ways}{a character string indicating how to merge the posterior two 
probability matrices obtained for each of the 2 databases. Four possibility are 
currently implemented: \code{"maxnorm"}, \code{"max"}, \code{"min"}, \code{"mean"} 
and \code{"prod"}. Default is \code{"maxnorm"}.}

\item{nb_perturb}{the number of perturbation used for the p-value combination.
Default is 1000.}

\item{outputfile_name}{a character string for the name of the .txt file where the 
results should be stored. Must be provided without extension. Default is 'Default_output'.}

\item{data}{a matrix of original (continuous) dataset, either a matrix or a dataframe.}
}
\value{
append the results in the file '\code{outputfile_name}.txt'.
}
\description{
P-values from matched probabilities
}
