% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mstep_C_sparse_big}
\alias{Mstep_C_sparse_big}
\title{C++ implementation of the E and M steps from Winkler's EM algorithm estimating FS method 
using sparse matrices for big sample sizes}
\usage{
Mstep_C_sparse_big(mat_A, mat_B, g_m, g_u, Log = FALSE)
}
\description{
C++ implementation of the E and M steps from Winkler's EM algorithm estimating FS method 
using sparse matrices for big sample sizes
}
\keyword{internal}

