% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_mix.R
\name{pdf_mix}
\alias{pdf_mix}
\title{Mixture of probability density functions}
\usage{
pdf_mix(x, pdf, w = NULL, parameters)
}
\arguments{
\item{x}{quantile where the probability density is to be evaluated.}

\item{pdf}{probability density function to be used that takes x as its first argument.}

\item{w}{a vector of lenght K containing the mixture component weights. Default is \code{NULL}, in which case equal 
proportions are assumed.}

\item{parameters}{a matrix of dimension K x p where p is the number of paramaters 
to be passed to the pdf function. The columns must be in the same order as the parameters}
}
\description{
Mixture of probability density functions
}

