% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perturb_FUN.R
\name{Perturb_FUN}
\alias{Perturb_FUN}
\title{Adding noise around data}
\usage{
Perturb_FUN(x, type = NULL, bet = 3)
}
\arguments{
\item{x}{original values}

\item{type}{a character string indicating which type of data is 'x'. Can be 
either binary (\code{"bin"} for binomial or \code{"bin_thresh"} for thresholding
a multivariate normal), continuous (\code{"con"}), or ... (\code{"dis"})
Default ids \code{"bin"}.}

\item{bet}{guessed value for the level of pertubation added}
}
\description{
Adding noise around data
}

