% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_zscore.R
\name{pval_zscore}
\alias{pval_zscore}
\title{Compute pvalues from Z-score}
\usage{
pval_zscore(beta, sigma)
}
\arguments{
\item{beta}{the estimate}

\item{sigma}{estimate's estimated variance}
}
\value{
the p-value
}
\description{
Compute pvalues from Z-score
}

