% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stringdate_test}
\alias{stringdate_test}
\title{Splitting a character string in C++}
\usage{
stringdate_test(Adates, i, l)
}
\arguments{
\item{Adates}{a matrix of character strings to be subsetted and splitted}

\item{i}{an integer indicating the begining of the subset of the strings}

\item{l}{an integer indicating the end of the subset of the strings}
}
\description{
Splitting a subset of each character string from character 
number \code{i} to character \code{l} on characters \code{";"}.
}
\examples{
stringdate_test(matrix(c("20NOV2008:21NOV2008", 
                        "20NOV2008:21NOV2008 ; 10APR2008:12APR2008", "", ""), 
                      ncol=2, nrow=2), 1,0)

}
\keyword{internal}

