% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_han2018.R
\name{test_han2018}
\alias{test_han2018}
\title{Association testing using Han & Lahiri estimating equations and jackknife approach}
\usage{
test_han2018(
  match_prob,
  y,
  x,
  jackknife_nrep = 100,
  methods = c("F", "M", "M2"),
  dist_family = c("gaussian", "binomial")
)
}
\arguments{
\item{match_prob}{matching probabilities matrix (e.g. obtained through \code{\link{recordLink}}) of 
dimensions \code{n1 x n2}.}

\item{y}{response variable of length \code{n1}. Only binary or gaussian phenotypes 
are supported at the moment.}

\item{x}{a \code{matrix} or a \code{data.frame} of predictors of dimensions \code{n2 x p}. 
An intercept is automatically added within the function.}

\item{jackknife_nrep}{the number of jackknife repetitions.
Default is 100.}

\item{methods}{a character vector which must be a subset of \code{("F", "M", "M2")} 
indicating which estimator from Han et al. 2018 should be computed. Default is all 3.}

\item{dist_family}{a character string indicating the distribution family for the glm. 
Currently, only \code{'gaussian'} and  \code{'binomial'} are supported. Default 
is \code{'gaussian'}.}
}
\value{
a list containing the following for each estimator in \code{methods}:
\itemize{
  \item \code{beta} a vector containing the \code{p} estimated coefficients
  \item \code{varcov} the \code{p x p} variance-covariance \code{matrix} of the \code{beta} coefficients
  \item \code{zscores} a vector containing the \code{p} Z-scores
  \item \code{pval} the corresponding Gaussian assumption p-values
}
}
\description{
Association testing using Han & Lahiri estimating equations and jackknife approach
}
\examples{
#rm(list=ls())
res <- list()
n_sims <- 1#5000
for(n in 1:n_sims){
x <- matrix(ncol=2, nrow=99, stats::rnorm(n=99*2))

#plot(density(rbeta(n=1000, 1,2)))
match_prob <- matrix(rbeta(n=103*99, 1, 2), nrow=103, ncol=99)


y <- rnorm(n=103, 1, 0.5)
res[[n]] <- test_han2018(match_prob, y, x)
#y <- rbinom(n=103, 1, prob=0.5)
#res[[n]] <- test_han2018(match_prob, y, x, dist_family = "binomial")
cat(n, "/", n_sims, "\n", sep="")
}
pvals <- sapply(lapply(res, "[[", "M2"), "[[", "pval")
rowMeans(pvals<0.05)

}
\references{
Han, Y., and Lahiri, P. (2019) Statistical Analysis with Linked Data. 
International Statistical Review, 87: S139– S157. https://doi.org/10.1111/insr.12295.
}
