% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludic-package.R
\docType{package}
\name{ludic-package}
\alias{ludic-package}
\alias{ludic}
\title{ludic}
\description{
Linkage Using Diagnosis Codes
}
\details{
This package implements probabilistic record linkage methods that relies on the use of diagnosis 
codes only, in the absence of direct identifiers .

\tabular{ll}{
Package: \tab ludic\cr
Type: \tab Package\cr
Version: \tab ludic 0.2.1\cr
Date: \tab 2025-07-24\cr
License: \tab \href{https://cran.r-project.org/web/licenses/MIT}{The "MIT License" (MIT)}\cr
}
The main function of \code{ludic} is \code{\link{recordLink}}.
}
\references{
Hejblum BP, Weber G, Liao KP, Palmer N, Churchill S, Szolovits P, 
Murphy S, Kohane I and Cai T, Probabilistic Record Linkage of De-Identified 
Research Datasets Using Diagnosis Codes, \emph{Scientific Data}, 6:180298 (2019). 
\doi{10.1038/sdata.2018.298}.

Zhang HG, Hejblum BP, Weber G, Palmer N, Churchill S, Szolovits P, 
Murphy S, Liao KP, Kohane I and Cai T, ATLAS: An automated association test using 
probabilistically linked health records with application to genetic studies, 
\emph{JAMIA}, in press (2021). 
\doi{10.1093/jamia/ocab187}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://borishejblum.github.io/ludic/}
  \item Report bugs at \url{https://github.com/borishejblum/ludic/issues}
}

}
\author{
Boris P. Hejblum, Harrison G. Zhang, Tianxi Cai
--- Maintainer: Boris P. Hejblum
}
