% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matchProbs_rank_full_C}
\alias{matchProbs_rank_full_C}
\title{Compute the matching probabilities for each pair of observations}
\usage{
matchProbs_rank_full_C(computed_dist, prop_match)
}
\arguments{
\item{computed_dist}{a \code{n x p} matrix of computed distances used for ranking.}

\item{prop_match}{a priori proportion of matches ("rho_1")}
}
\value{
a \code{n x p} matrix containing the matching probabilities for each pair
}
\description{
C++ version: for each observations in \code{(1:n)}, all the matching probabilities are computed
for the \code{p} possible pairs.
}
