% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordLink.R
\name{recordLink}
\alias{recordLink}
\title{Probabilistic Patient Record Linkage}
\usage{
recordLink(data1, data2, dates1 = NULL, dates2 = NULL, eps_plus, eps_minus,
  aggreg_2ways = "mean", min_prev = 0.01, data1_cont2diff, data2_cont2diff,
  eps_inf, d_max, use_diff = TRUE)
}
\arguments{
\item{data1}{either a binary matrix or dataframe whose rownames are .}

\item{data2}{either a binary matrix or a dataframe whose rownames are .}

\item{eps_plus}{discrepancy rate between \code{data1} and \code{data2}}

\item{eps_minus}{discrepancy rate between \code{data2} and \code{data1}}

\item{aggreg_2ways}{a charatcer string indicating how to merge the posterior two 
probability matrices obtained for each of the 2 databases. Four possibility are 
currently implemented: \code{"maxnorm"}, \code{"max"}, \code{"min"}, \code{"mean"} 
and \code{"prod"}. Default is \code{"mean"}.}

\item{min_prev}{minimum prevalence for the variables used in matching.
Default is 1\%.}

\item{data1_cont2diff}{either a matrix or dataframe of continuous features, 
such as age, for which the similarity measure uses the difference with 
\code{data2_cont2diff}, whose rownames are .}

\item{data2_cont2diff}{either a matrix or dataframe of continuous features, 
such as age, for which the similarity measure uses the difference with 
\code{data2_cont1diff}, whose rownames are .}
}
\value{
the posterior probability of matching matrix
}
\description{
Probabilistic Patient Record Linkage
}

