% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{strsplitC}
\alias{strsplitC}
\title{Splitting a character string in C++}
\usage{
strsplitC(s, sep)
}
\arguments{
\item{s}{a character string to be split}

\item{sep}{a character that delimits the splits}
}
\description{
Splitting a character string in C++
}
\examples{
strsplitC(c(";aaa;bb;cccc;ee;"), sep=";")
}
\keyword{internal}
